Factor out the suggested methods, added literature references and combined my clases implemented into one
