+++ jgrapht-core/src/main/java/org/jgrapht/alg/shortestpath/LabelCorrectingMultiObjectiveShortestPath.java
@@ -19,6 +19,7 @@ package org.jgrapht.alg.shortestpath;
 
 import org.jgrapht.*;
 import org.jgrapht.graph.*;
+import org.jgrapht.util.*;
 
 import java.util.*;
 import java.util.function.*;
@@ -30,31 +31,51 @@ import java.util.stream.*;
  * <p>
  * The algorithm is a multiple objective extension of the Bellman-Ford relaxation process. It
  * maintains a set of non-dominated labels for each vertex and iteratively propagates labels through
- * outgoing edges. A newly generated label is discarded if it is dominated by an existing label at the
- * same vertex. Existing labels dominated by the new label are removed.
+ * outgoing edges. A newly generated label is discarded if it is dominated by an existing label at
+ * the same vertex. Existing labels dominated by the new label are removed.
+ *
+ * <p>
+ * Optionally, an approximation tolerance $\varepsilon$ may be provided. In that case labels which
+ * are sufficiently close according to the tolerance are also discarded in order to reduce the
+ * number of stored labels, and the algorithm computes an approximation of the Pareto set. With a
+ * tolerance equal to zero the complete Pareto set is computed.
  *
  * <p>
  * All objective values must be non-negative.
+ *
+ * <p>
+ * The label-correcting strategy is described in: A. J. V. Skriver and K. A. Andersen. (2000). A
+ * label correcting approach for solving bicriterion shortest-path problems. Computers &amp;
+ * Operations Research. 27. 507-524. 10.1016/S0305-0548(99)00037-4.
+ *
+ * <p>
+ * The approximation of the Pareto set is described in: A. Warburton. (1987). Approximation of
+ * Pareto Optima in Multiple-Objective, Shortest-Path Problems. Operations Research. 35. 70-79.
+ * 10.1287/opre.35.1.70.
+ *
+ * <p>
+ * Note that the multi-objective shortest path problem is a well-known NP-hard problem.
  *
  * @author Mario Fuentes Jimenez
  *
  * @param <V> the vertex type
  * @param <E> the edge type
  */
-public class LabelCorrectingMultiObjectiveShortestPath<V, E>
-    extends BaseMultiObjectiveShortestPathAlgorithm<V, E>
+public class LabelCorrectingMultiObjectiveShortestPath<V, E> extends BaseMultiObjectiveShortestPathAlgorithm<V, E>
 {
     // the edge weight function
     private final Function<E, double[]> edgeWeightFunction;
     // the number of objectives
     private final int objectives;
+    // the approximation tolerance
+    private final double epsilon;
     // final labels for each node
     private final Map<V, LinkedList<Label>> nodeLabels;
     // labels waiting to be expanded
     private final Queue<Label> queue;
 
     /**
-     * Create a new shortest path algorithm.
+     * Create a new shortest path algorithm which computes the complete Pareto set.
      *
      * @param graph the input graph
      * @param edgeWeightFunction the edge weight function
@@ -62,27 +83,38 @@ public class LabelCorrectingMultiObjectiveShortestPath<V, E>
     public LabelCorrectingMultiObjectiveShortestPath(
         Graph<V, E> graph, Function<E, double[]> edgeWeightFunction)
     {
+        this(graph, edgeWeightFunction, 0d);
+    }
+
+    /**
+     * Create a new shortest path algorithm which computes an epsilon-approximation of the Pareto
+     * set. A tolerance equal to zero computes the complete Pareto set.
+     *
+     * @param graph the input graph
+     * @param edgeWeightFunction the edge weight function
+     * @param epsilon the approximation tolerance, must be non-negative
+     */
+    public LabelCorrectingMultiObjectiveShortestPath(
+        Graph<V, E> graph, Function<E, double[]> edgeWeightFunction, double epsilon)
+    {
         super(graph);
         this.edgeWeightFunction =
             Objects.requireNonNull(edgeWeightFunction, "Function cannot be null");
+        if (Double.compare(epsilon, 0d) < 0) {
+            throw new IllegalArgumentException("Epsilon must be non-negative");
+        }
         this.objectives = validateEdgeWeightFunction(edgeWeightFunction);
+        this.epsilon = epsilon;
         this.nodeLabels = new HashMap<>();
         this.queue = new ArrayDeque<>();
     }
 
-
-    /**
-    * {@inheritDoc}
-    */
     @Override
     public List<GraphPath<V, E>> getPaths(V source, V sink)
     {
         return this.getPaths(source).getPaths(sink);
     }
 
-    /**
-    * {@inheritDoc}
-    */
     @Override
     public MultiObjectiveSingleSourcePaths<V, E> getPaths(V source)
     {
@@ -126,25 +158,30 @@ public class LabelCorrectingMultiObjectiveShortestPath<V, E>
 
             for (E e : graph.outgoingEdgesOf(v)) {
                 V u = Graphs.getOppositeVertex(graph, e, v);
-                Label newLabel =
-                    new Label(u, sum(curLabel.value, edgeWeightFunction.apply(e)), curLabel, e);
+                Label newLabel = new Label(
+                    u, MathUtil.vectorSum(curLabel.value, edgeWeightFunction.apply(e)), curLabel,
+                    e);
 
-                boolean isDominated = false;
+                boolean discard = false;
                 LinkedList<Label> uLabels = nodeLabels.get(u);
                 ListIterator<Label> it = uLabels.listIterator();
 
                 while (it.hasNext()) {
                     Label oldLabel = it.next();
-                    if (dominates(oldLabel.value, newLabel.value)) {
-                        isDominated = true;
+                    if (MathUtil.vectorDominates(oldLabel.value, newLabel.value)) {
+                        discard = true;
                         break;
                     }
-                    if (dominates(newLabel.value, oldLabel.value)) {
+                    if (epsilon > 0d && close(oldLabel.value, newLabel.value, epsilon)) {
+                        discard = true;
+                        break;
+                    }
+                    if (MathUtil.vectorDominates(newLabel.value, oldLabel.value)) {
                         it.remove();
                     }
                 }
 
-                if (!isDominated) {
+                if (!discard) {
                     uLabels.add(newLabel);
                     queue.add(newLabel);
                 }
@@ -182,72 +219,23 @@ public class LabelCorrectingMultiObjectiveShortestPath<V, E>
     }
 
     /**
-     * Compute the sum of two vectors.
+     * Return whether two vectors are close according to the epsilon tolerance.
      *
      * @param a the first vector
      * @param b the second vector
-     * @return the sum
+     * @param epsilon the approximation tolerance
+     * @return true if the vectors are close
      */
-    private static double[] sum(double[] a, double b[])
+    private static boolean close(double[] a, double[] b, double epsilon)
     {
         int d = a.length;
-        double[] res = new double[d];
         for (int i = 0; i < d; i++) {
-            res[i] = a[i] + b[i];
-        }
-        return res;
-    }
-
-    /**
-     * Return whether a vector dominates another.
-     *
-     * @param a the first vector
-     * @param b the second vector
-     * @return true if the first vector dominates the second
-     */
-    private static boolean dominates(double[] a, double[] b)
-    {
-        boolean strict = false;
-        int d = a.length;
-        for (int i = 0; i < d; i++) {
-            if (a[i] > b[i]) {
+            double tolerance = Math.abs(a[i]) * epsilon;
+            if (Math.abs(a[i] - b[i]) > tolerance) {
                 return false;
             }
-            if (a[i] < b[i]) {
-                strict = true;
-            }
         }
-        return strict;
-    }
-
-    /**
-     * Check the validity of the edge weight function.
-     *
-     * @param edgeWeightFunction the edge weight function
-     * @return the number of dimensions
-     */
-    private int validateEdgeWeightFunction(Function<E, double[]> edgeWeightFunction)
-    {
-        int dim = 0;
-        for (E e : graph.edgeSet()) {
-            double[] f = edgeWeightFunction.apply(e);
-            if (f == null) {
-                throw new IllegalArgumentException("Invalid edge weight function");
-            }
-            if (dim == 0) {
-                dim = f.length;
-            } else {
-                if (dim != f.length) {
-                    throw new IllegalArgumentException("Invalid edge weight function");
-                }
-            }
-            for (int i = 0; i < dim; i++) {
-                if (Double.compare(f[i], 0d) < 0) {
-                    throw new IllegalArgumentException("Edge weight must be non-negative");
-                }
-            }
-        }
-        return dim;
+        return true;
     }
 
     /**
